+++ transport/http2_client.go
@@ -1220,7 +1220,7 @@ func (t *http2Client) ChannelzMetric() *channelz.SocketInternalMetric {
 		RemoteAddr:                      t.remoteAddr,
 		// RemoteName :
 	}
-	if au, ok := t.authInfo.(credentials.Security); ok {
+	if au, ok := t.authInfo.(credentials.ExtraSecurityInfo); ok {
 		s.Security = au.GetSecurityValue()
 	}
 	t.czmu.RUnlock()
